refactor(client): split ResourceModulesAdmin into focused sections
Decompose the top complexity hotspot (issue #453). The component owned
group/module/log UI state plus a 242-line inline groups.map that branched
on edit/create states.

- Lift the UI state machine into useModuleAdminUiState
- Extract route-private section components: ModuleAdminHeader,
  ModuleAssistDialog, UngroupedModulesSection, ModuleGroupSection (with
  local GroupTitle/GroupMeta/GroupHeaderControls/GroupModuleList/
  GroupProgressHint), and a shared ModuleListItem
- Expose ResourceModulesController type so sections compose over the hook

ResourceModulesAdmin is now a thin composition; every unit is below
cyc 10 / cog 15. No behavior change.
